CODERUB: Updated dotnet.yml withi

+++ FlutterWave.Core/Models/Services/Foundations/FlutterWave/Transfers/AllTransfersResponse.cs
@@ -38,7 +38,7 @@ namespace FlutterWave.Core.Models.Services.Foundations.FlutterWave.Transfers
             public string Currency { get; set; }
             public string DebitCurrency { get; set; }
             public int Amount { get; set; }
-            public int Fee { get; set; }
+            public double Fee { get; set; }
             public string Status { get; set; }
             public string Reference { get; set; }
             public object Meta { get; set; }
